Concatenates files with Ruby and removed variables_for_type architecture.
In an attempt to increase cross-platform compatibility, I have removed the
dependency on the `cat` command-line utility in favour of Ruby's IO capabilities.
Gone are the variables_for_javascript and variables_for_stylesheet in favour of
more detailed javascript_include_compressed and stylesheet_include_compressed
methods which, in turn, use include_compressed's improved interface.

`include_compressed` now takes an asset type's tag helper as a Proc object as well
as the compressed file name and sources which are pre-formatted to be relative
from RAILS_ROOT/public.

A minor cosmetic change was to switch from using the Proc array notation to
the slightly less confusing `call` method.